Create MultiSigWallet.sol
This contract allows the owner to add and remove signatory wallets, and requires a minimum of 60% of the signatory wallets to sign a transaction before it is executed. It uses the sign() function to get signatures from the signatory wallets, which is a built-in function in Solidity that returns a signature that can be verified by the transfer() function.